+++ controlador/parser/binario.cpp
@@ -5,7 +5,7 @@ namespace expre{
     ** Parâmetros: O valor da variável
     ** Retornos: A soma das expressões
     */
-    float soma::calcular (const float &x) {
+    double soma::calcular (const double &x) {
         return filhoEsq->calcular(x) + filhoDir->calcular(x);
     }
 
@@ -13,7 +13,7 @@ namespace expre{
     ** Parâmetros: O valor da variável
     ** Retornos: A subtração das expressões
     */
-    float subtracao::calcular (const float &x) {
+    double subtracao::calcular (const double &x) {
         return filhoEsq->calcular(x) - filhoDir->calcular(x);
     }
 
@@ -21,7 +21,7 @@ namespace expre{
     ** Parâmetros: O valor da variável
     ** Retornos: A multiplicação das expressões
     */
-    float multiplicacao::calcular (const float &x) {
+    double multiplicacao::calcular (const double &x) {
         return filhoEsq->calcular(x) * filhoDir->calcular(x);
     }
 
@@ -29,7 +29,7 @@ namespace expre{
     ** Parâmetros: O valor da variável
     ** Retornos: A divisão das expressões
     */
-    float divisao::calcular (const float &x) {
+    double divisao::calcular (const double &x) {
         return filhoEsq->calcular(x) / filhoDir->calcular(x);
     }
 
@@ -37,7 +37,7 @@ namespace expre{
     ** Parâmetros: O valor da variável
     ** Retornos: A potência das expressões
     */
-    float potencia::calcular (const float &x) {
+    double potencia::calcular (const double &x) {
         return std::pow(filhoEsq->calcular(x), filhoDir->calcular(x));
     }
 }
